+++ DiscoveryClassifier.ServiceClient/DiscoveryClassifier.ServiceClient.csproj
@@ -64,13 +64,14 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="CategoryServiceClient.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="RESTParameter\CategoryCheckRequest.cs" />
     <Compile Include="RESTParameter\ErrorResponse.cs" />
     <Compile Include="RESTParameter\PublishRequest.cs" />
     <Compile Include="RESTParameter\SearchRequest.cs" />
     <Compile Include="RESTServiceClient.cs" />
+    <Compile Include="RESTServiceClientBase.cs" />
+    <Compile Include="RESTServiceClientCategories.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\DiscoveryClassifier.BusinessObjects\DiscoveryClassifier.BusinessObjects.csproj">
